rename to scraper_page pattern
