+++ scraper_mythic.py
